+++ frontend/src/components/Item.jsx
@@ -1,0 +1,19 @@
+import React, { Component } from "react";
+
+import { ListToggle } from "./ListToggle";
+
+export class Item extends Component {
+  render() {
+    return (
+      <div className="Item">
+        <div className="ItemContainer" style={{ backgroundImage: `url(${this.props.backdrop})` }}>
+          <div className="overlay">
+            <div className="title">{this.props.title}</div>
+            <div className="rating">{this.props.score} / 10</div>
+            <ListToggle />
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
