+++ frontend/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-import App from './App';
+import App from './components/App';
 import './index.css';
 
 const root = createRoot(document.getElementById('root'));
